test: split unit tests by inheritance hierarchy chore: use "read" instead of "getter" chore: use "write" instead of "effects"
