+++ test/unit/strategies/uniswap-v2/uniswap-v2.behavior.ts
@@ -1,0 +1,9 @@
+import { shouldBehaveLikeLiquidate } from "./write/liquidate";
+
+export function shouldBehaveLikeUniswapV2Strategy(): void {
+  describe("Write Methods", function () {
+    describe("liquidate", function () {
+      shouldBehaveLikeLiquidate();
+    });
+  });
+}
