+++ test/unit/strategies/uniswap-v2/write/liquidate.ts
@@ -1,0 +1,15 @@
+import { expect } from "chai";
+
+export function shouldBehaveLikeLiquidate(): void {
+  context("when the pair does not exist", function () {
+    beforeEach(async function () {});
+
+    it("reverts", async function () {});
+  });
+
+  context("when the pair exists", function () {
+    beforeEach(async function () {});
+
+    it("liquidates the underwater vaults", async function () {});
+  });
+}
