+++ test/unit/strategies/uniswap-v3/uniswap-v3.behavior.ts
@@ -1,0 +1,9 @@
+import { shouldBehaveLikeLiquidate } from "./write/liquidate";
+
+export function shouldBehaveLikeUniswapV3Strategy(): void {
+  describe("Write Methods", function () {
+    describe("liquidate", function () {
+      shouldBehaveLikeLiquidate();
+    });
+  });
+}
