+++ test/unit/strategies/uniswap-v3/write/liquidate.ts
@@ -1,0 +1,7 @@
+import { expect } from "chai";
+
+export function shouldBehaveLikeLiquidate(): void {
+  beforeEach(async function () {});
+
+  it("liquidates the underwater vaults", async function () {});
+}
